+++ doc/change_log.md
@@ -1186,3 +1186,9 @@ script:
   `echo "$exec_t_comp_ratioB + $ratioB"|bc` to 
   echo $exec_t_comp_ratioB $ratioB | awk '{ printf "%0.2f\n", $1+$2}
   to aviod issue: 0.25 is shown as .25
+
+## 20240115
+
+finish the memory planner, with two mode:
+  1. priority mode (global)
+  2. sequential mode